TFP-2974 Bruker Beregningsresulat i stedet for Uttak for å hente informasjon om første og siste uttaksdato(utbetalingsdato) når vi publiserer vedtakshendelser for SVP og FP. Når vi henter informasjonen fra uttak får vi med perioder som er utsatt, eller perioder som er avslått og det ønsker vi ikke (gir falsk overlapp).
